new helper file to house tabbing logic, applied to fish, supp, procedure, and telemetry
